ingest/requirements.txt
psycopg2-binary>=2.9.0
watchdog>=4.0.0
pyyaml>=6.0
